Разделяет компонент days на два: days.js и day.js
